Add AbstractNewFolderWizardPagec class
